+++ FileHelpers.Examples/FileHelpers.Examples.csproj
@@ -79,9 +79,9 @@
     <Compile Include="Examples\18.Converters\50.EnumConverter.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Examples\25.ErrorHandling\10.ErrorHandling.cs" />
-    <Compile Include="Examples\25.ErrorHandling\30.ErrorHandlingIgnoreMode.cs" />
-    <Compile Include="Examples\25.ErrorHandling\40.ErrorHandlingErrorModeSave.cs" />
+    <Compile Include="Examples\25.ErrorHandling\10.ErrorMode.ThrowException.cs" />
+    <Compile Include="Examples\25.ErrorHandling\30.ErrorMode.IgnoreAndContinue.cs" />
+    <Compile Include="Examples\25.ErrorHandling\40.ErrorMode.SaveAndContinue.cs" />
     <Compile Include="Examples\50.Advanced\10.MultipleDelimiters.cs" />
     <Compile Include="Examples\50.Advanced\30.MultiRecordSample.cs" />
     <Compile Include="Examples\50.Advanced\05.DynamicChangeOptions.cs" />
